templates/telegram-bot-template/requirements.txt
python-telegram-bot==20.7
requests==2.31.0
python-dotenv==1.0.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
passlib[bcrypt]==1.7.4
python-jose==3.3.0
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.0
